+++ test/server/http/admin_test.cc
@@ -53,7 +53,7 @@ public:
 
   static std::string
   statsAsJsonHandler(std::map<std::string, uint64_t>& all_stats,
-                     const std::list<Stats::ParentHistogramSharedPtr>& all_histograms) {
+                     const std::vector<Stats::ParentHistogramSharedPtr>& all_histograms) {
     return AdminImpl::statsAsJson(all_stats, all_histograms, true);
   }
 
@@ -580,8 +580,8 @@ TEST(PrometheusStatsFormatter, MetricNameCollison) {
   //`statsAsPrometheus()` should return two implying it found two unique stat names
 
   Stats::HeapRawStatDataAllocator alloc;
-  std::list<Stats::CounterSharedPtr> counters;
-  std::list<Stats::GaugeSharedPtr> gauges;
+  std::vector<Stats::CounterSharedPtr> counters;
+  std::vector<Stats::GaugeSharedPtr> gauges;
 
   {
     std::string name = "cluster.test_cluster_1.upstream_cx_total";
@@ -638,8 +638,8 @@ TEST(PrometheusStatsFormatter, UniqueMetricName) {
   // four unique stat names.
 
   Stats::HeapRawStatDataAllocator alloc;
-  std::list<Stats::CounterSharedPtr> counters;
-  std::list<Stats::GaugeSharedPtr> gauges;
+  std::vector<Stats::CounterSharedPtr> counters;
+  std::vector<Stats::GaugeSharedPtr> gauges;
 
   {
     std::string name = "cluster.test_cluster_1.upstream_cx_total";